[Fix #160295024] Restructures the app Restructures api into v1 folder for challenge 2
